use powf/logf/expf
seems pow/log/exp can't be exported for libc 2.27 or above so use float versions which are exported apparently

+++ src/addrman.cpp
@@ -60,7 +60,7 @@ bool CAddrInfo::IsTerrible(int64_t nNow) const
 
 double CAddrInfo::GetChance(int64_t nNow) const
 {
-    double fChance = 1.0;
+    float fChance = 1.0;
     int64_t nSinceLastTry = std::max<int64_t>(nNow - nLastTry, 0);
 
     // deprioritize very recent attempts away
@@ -68,9 +68,9 @@ double CAddrInfo::GetChance(int64_t nNow) const
         fChance *= 0.01;
 
     // deprioritize 66% after each failed attempt, but at most 1/28th to avoid the search taking forever or overly penalizing outages.
-    fChance *= pow(0.66, std::min(nAttempts, 8));
+    fChance *= powf(0.66, std::min(nAttempts, 8));
 
-    return fChance;
+    return (double)fChance;
 }
 // SYSCOIN
 CAddrInfo* CAddrMan::Find(const CService& addr, int* pnId)
